Update typescript and related dependencies
We're trying to extend the validator dir in tf-cdn-iac but we're
running into dependency conflicts which are due to the older version of
Typescript pinned by the firewalker package. To unblock the work we're
doing in tf-cdn-iac, we need to update the typescript version used here
and the related packages that depend on a particular typescript
version.

I've confirmed that these changes should be safe by running both
`yarn build` and `yarn test`.

There were some conflicts related to type definitions for ffi-related
packages, hence the introduction of a `types` folder to resolve these
issues.